Uses ArrayList to create inventory
To address the possibility of a player owning
multiple items, an array was made in order 
to keep track of these items. Use array instead
of declaring multiple member items that may not 
be enough or may be too much. Here the Array 
expands to accommodate just as much data as is
needed.